// src/generated/Protos/V1/Models/InferenceModels.ts
export enum InferenceCloseReasons {
  CloseTimeout = 0,
  CloseHostClosing = 1,
  CloseError = 2,
  CloseRequestedByClient = 3,
  UNRECOGNIZED = -1,
}
// ...
export enum ThinkLevels {
  ThinkLevelsBasic = 0,
  ThinkLevelsBasicWithTools = 1,
  ThinkLevelsSkilled = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum InferenceToolGroups {
  InferenceToolGroupsInformationTools = 0,
  InferenceToolGroupsFileTools = 1,
  InferenceToolGroupsMathTools = 2,
  InferenceToolGroupsCommunicationTools = 3,
  InferenceToolGroupsCodingTools = 4,
  InferenceToolGroupsMediaTools = 5,
  InferenceToolGroupsIntegrationTools = 6,
  InferenceToolGroupsSocialTools = 7,
  InferenceToolGroupsShellTools = 8,
  InferenceToolGroupsScreenTools = 9,
  InferenceToolGroupsInputTools = 10,
  InferenceToolGroupsClipboardTools = 11,
  InferenceToolGroupsBrowserTools = 12,
  InferenceToolGroupsWindowTools = 13,
  InferenceToolGroupsSystemTools = 14,
  InferenceToolGroupsGitTools = 15,
  UNRECOGNIZED = -1,
}
// ...
export interface SendInferenceRequest {
  SessionId: string;
  InferenceId: string;
  Text: string;
  ToolGroup?: InferenceToolGroups | undefined;
  ToolName?: string | undefined;
  AntiPrompts: string[];
  TokensKeep?: number | undefined;
  MaxTokens?: number | undefined;
  DecodeSpecialTokens?: boolean | undefined;
  Temperature?: number | undefined;
  TopP?: number | undefined;
  TopK?: number | undefined;
  RepeatPenalty?: number | undefined;
  Seed?: number | undefined;
  FrequencyPenalty?: number | undefined;
  PresencePenalty?: number | undefined;
  MinKeep?: number | undefined;
  MinP?: number | undefined;
  PenalizeNewline?: boolean | undefined;
  PenaltyCount?: number | undefined;
  PreventEOS?: boolean | undefined;
  TypicalP?: number | undefined;
  ThinkLevel?: ThinkLevels | undefined;
  OutputFormat?: InferenceOutputFormats | undefined;
  ExampleOutput?: string | undefined;
}

export interface SendInferenceResponse {
  SessionId: string;
  InferenceId: string;
  Id: string;
  Type: InferenceResponseTypes;
  Content: string;
  AuthorRole: string;
  Format: InferenceOutputFormats;
  MessageTokenCount: number;
  SessionTokenCount: number;
  ComputeTimeMs: number;
}

export interface InferenceStatsRequest {
  SessionId: string;
  InferenceId: string;
}

export interface InferenceStatsResponse {
  SessionId: string;
  InferenceId: string;
  Success: boolean;
  LastMessageTokenCount: number;
  SessionTokenCount: number;
  LastMessageToolCount: number;
  SessionToolCount: number;
  LastMessageComputeTimeMs: number;
  SessionComputeTimeMs: number;
}

export interface CreateInferenceRequest {
  SessionId: string;
  ModelName: string;
  InitializationPrompt: string;
  ThinkLevel: ThinkLevels;
  ToolGroups: InferenceToolGroups[];
}

export interface CreateInferenceResponse {
  SessionId: string;
  InferenceId: string;
}

export interface CloseInferenceRequest {
  SessionId: string;
  InferenceId: string;
  Reason: InferenceCloseReasons;
}

export interface CloseInferenceResponse {
  SessionId: string;
  InferenceId: string;
  Success: boolean;
}
// ...
export const InferenceStatsRequest: MessageFns<InferenceStatsRequest> = {
  encode(message: InferenceStatsRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.SessionId !== "") {
      writer.uint32(10).string(message.SessionId);
    }
    if (message.InferenceId !== "") {
      writer.uint32(18).string(message.InferenceId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): InferenceStatsRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseInferenceStatsRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.SessionId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.InferenceId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): InferenceStatsRequest {
    return {
      SessionId: isSet(object.SessionId) ? globalThis.String(object.SessionId) : "",
      InferenceId: isSet(object.InferenceId) ? globalThis.String(object.InferenceId) : "",
    };
  },

  toJSON(message: InferenceStatsRequest): unknown {
    const obj: any = {};
    if (message.SessionId !== "") {
      obj.SessionId = message.SessionId;
    }
    if (message.InferenceId !== "") {
      obj.InferenceId = message.InferenceId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<InferenceStatsRequest>, I>>(base?: I): InferenceStatsRequest {
    return InferenceStatsRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<InferenceStatsRequest>, I>>(object: I): InferenceStatsRequest {
    const message = createBaseInferenceStatsRequest();
    message.SessionId = object.SessionId ?? "";
    message.InferenceId = object.InferenceId ?? "";
    return message;
  },
};
// ...
export const CreateInferenceRequest: MessageFns<CreateInferenceRequest> = {
  encode(message: CreateInferenceRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.SessionId !== "") {
      writer.uint32(10).string(message.SessionId);
    }
    if (message.ModelName !== "") {
      writer.uint32(18).string(message.ModelName);
    }
    if (message.InitializationPrompt !== "") {
      writer.uint32(26).string(message.InitializationPrompt);
    }
    if (message.ThinkLevel !== 0) {
      writer.uint32(32).int32(message.ThinkLevel);
    }
    writer.uint32(42).fork();
    for (const v of message.ToolGroups) {
      writer.int32(v);
    }
    writer.join();
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateInferenceRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateInferenceRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.SessionId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.ModelName = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.InitializationPrompt = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.ThinkLevel = reader.int32() as any;
          continue;
        }
        case 5: {
          if (tag === 40) {
            message.ToolGroups.push(reader.int32() as any);

            continue;
          }

          if (tag === 42) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.ToolGroups.push(reader.int32() as any);
            }

            continue;
          }

          break;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateInferenceRequest {
    return {
      SessionId: isSet(object.SessionId) ? globalThis.String(object.SessionId) : "",
      ModelName: isSet(object.ModelName) ? globalThis.String(object.ModelName) : "",
      InitializationPrompt: isSet(object.InitializationPrompt) ? globalThis.String(object.InitializationPrompt) : "",
      ThinkLevel: isSet(object.ThinkLevel) ? thinkLevelsFromJSON(object.ThinkLevel) : 0,
      ToolGroups: globalThis.Array.isArray(object?.ToolGroups)
        ? object.ToolGroups.map((e: any) => inferenceToolGroupsFromJSON(e))
        : [],
    };
  },

  toJSON(message: CreateInferenceRequest): unknown {
    const obj: any = {};
    if (message.SessionId !== "") {
      obj.SessionId = message.SessionId;
    }
    if (message.ModelName !== "") {
      obj.ModelName = message.ModelName;
    }
    if (message.InitializationPrompt !== "") {
      obj.InitializationPrompt = message.InitializationPrompt;
    }
    if (message.ThinkLevel !== 0) {
      obj.ThinkLevel = thinkLevelsToJSON(message.ThinkLevel);
    }
    if (message.ToolGroups?.length) {
      obj.ToolGroups = message.ToolGroups.map((e) => inferenceToolGroupsToJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateInferenceRequest>, I>>(base?: I): CreateInferenceRequest {
    return CreateInferenceRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateInferenceRequest>, I>>(object: I): CreateInferenceRequest {
    const message = createBaseCreateInferenceRequest();
    message.SessionId = object.SessionId ?? "";
    message.ModelName = object.ModelName ?? "";
    message.InitializationPrompt = object.InitializationPrompt ?? "";
    message.ThinkLevel = object.ThinkLevel ?? 0;
    message.ToolGroups = object.ToolGroups?.map((e) => e) || [];
    return message;
  },
};

function createBaseCreateInferenceResponse(): CreateInferenceResponse {
  return { SessionId: "", InferenceId: "" };
}

export const CreateInferenceResponse: MessageFns<CreateInferenceResponse> = {
  encode(message: CreateInferenceResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.SessionId !== "") {
      writer.uint32(10).string(message.SessionId);
    }
    if (message.InferenceId !== "") {
      writer.uint32(18).string(message.InferenceId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateInferenceResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateInferenceResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.SessionId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.InferenceId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CreateInferenceResponse {
    return {
      SessionId: isSet(object.SessionId) ? globalThis.String(object.SessionId) : "",
      InferenceId: isSet(object.InferenceId) ? globalThis.String(object.InferenceId) : "",
    };
  },

  toJSON(message: CreateInferenceResponse): unknown {
    const obj: any = {};
    if (message.SessionId !== "") {
      obj.SessionId = message.SessionId;
    }
    if (message.InferenceId !== "") {
      obj.InferenceId = message.InferenceId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CreateInferenceResponse>, I>>(base?: I): CreateInferenceResponse {
    return CreateInferenceResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CreateInferenceResponse>, I>>(object: I): CreateInferenceResponse {
    const message = createBaseCreateInferenceResponse();
    message.SessionId = object.SessionId ?? "";
    message.InferenceId = object.InferenceId ?? "";
    return message;
  },
};
// ...
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}